requirements.txt
pandas==2.0.3
openpyxl==3.1.2
moviepy==1.0.3
Pillow==10.0.0
google-api-python-client==2.95.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
schedule==1.2.0
pydub==0.25.1
requests==2.31.0
python-dotenv==1.0.0